avoid concurrent sub/unsub on same channel (#26454)

+++ internal/querynodev2/delegator/delegator_data.go
@@ -250,6 +250,10 @@ func (sd *shardDelegator) applyDelete(ctx context.Context, nodeID int64, worker 
 		if ok {
 			log.Debug("delegator plan to applyDelete via worker")
 			err := retry.Do(ctx, func() error {
+				if sd.Stopped() {
+					return retry.Unrecoverable(merr.WrapErrChannelUnsubscribing(sd.vchannelName))
+				}
+
 				err := worker.Delete(ctx, &querypb.DeleteRequest{
 					Base:         commonpbutil.NewMsgBase(commonpbutil.WithTargetID(nodeID)),
 					CollectionId: sd.collectionID,
